feat: add result format versioning
BREAKING CHANGE: but not really. Marking this as a breaking change in
order to more carefully stage this code as a second line of defense. In
theory, there should be no publicly visible breaking changes. However,
because we might be messing with existing on-disk data, it is worth
following opt-in semantics.